Move all tab components to group view dir
