--- frontend/src/views/groups/tabs/index.tsx
@@ -1,3 +1,0 @@
-import { Tabs } from "./Tabs";
-
-export { Tabs };
