--- frontend/src/views/groups/tabs/modal/createPunishmentTableRow/index.tsx
@@ -1,3 +1,0 @@
-import { CreatePunishmentTableRow } from "./CreatePunishmentTableRow";
-
-export { CreatePunishmentTableRow };
